fix: update memory command import paths after cli.py extraction
AgentRegistry and A2AClient were accessed via cli_module indirection.
Import them directly and fix test mock paths accordingly.
